requirements.txt
flask>=2.3,<3.0
click>=8.1.3,<9.0
itsdangerous>=2.1.2,<3.0
Jinja2>=3.1.2,<4.0
MarkupSafe>=2.1.2,<3.0
Werkzeug>=2.3.3,<3.0
pymongo
flask-httpauth
flask_restful
apscheduler
requests
simple-file-checksum
python-magic